Extract real into its own module; add reals![...] macro
