+++ godot-core/src/builtin/mod.rs
@@ -33,7 +33,7 @@
 //!   overloading would become impossible](https://github.com/kvark/mint/issues/75).
 
 // Re-export macros.
-pub use crate::{array, dict, varray};
+pub use crate::{array, dict, real, reals, varray};
 
 pub use aabb::*;
 pub use array_inner::{Array, VariantArray};
@@ -47,6 +47,7 @@ pub use packed_array::*;
 pub use plane::*;
 pub use projection::*;
 pub use quaternion::*;
+pub use real_inner::*;
 pub use rect2::*;
 pub use rect2i::*;
 pub use rid::*;
@@ -75,12 +76,7 @@ pub mod dictionary {
 // Modules exporting declarative macros must appear first.
 mod macros;
 
-// Rename imports because we re-export a subset of types under same module names.
-#[path = "array.rs"]
-mod array_inner;
-#[path = "dictionary.rs"]
-mod dictionary_inner;
-
+// Other modules
 mod aabb;
 mod basis;
 mod callable;
@@ -100,6 +96,17 @@ mod transform2d;
 mod transform3d;
 mod variant;
 mod vectors;
+
+// Rename imports because we re-export a subset of types under same module names.
+#[path = "array.rs"]
+mod array_inner;
+#[path = "dictionary.rs"]
+mod dictionary_inner;
+#[path = "real.rs"]
+mod real_inner;
+
+// Glam re-exports
+pub(crate) use glam::{IVec2, IVec3, IVec4};
 
 #[doc(hidden)]
 pub mod inner {
@@ -124,202 +131,6 @@ pub(crate) fn u8_to_bool(u: u8) -> bool {
         1 => true,
         _ => panic!("Invalid boolean value {u}"),
     }
-}
-
-/// Clippy often complains if you do `f as f64` when `f` is already an `f64`. This trait exists to make it easy to
-/// convert between the different reals and floats without a lot of allowing clippy lints for your code.
-pub trait RealConv {
-    /// Cast this [`real`][type@real] to an [`f32`] using `as`.
-    // Clippy complains that this is an `as_*` function but it takes a `self`
-    // however, since this uses `as` internally it makes much more sense for
-    // it to be named `as_f32` rather than `to_f32`.
-    #[allow(clippy::wrong_self_convention)]
-    fn as_f32(self) -> f32;
-
-    /// Cast this [`real`][type@real] to an [`f64`] using `as`.
-    // Clippy complains that this is an `as_*` function but it takes a `self`
-    // however, since this uses `as` internally it makes much more sense for
-    // it to be named `as_f64` rather than `to_f64`.
-    #[allow(clippy::wrong_self_convention)]
-    fn as_f64(self) -> f64;
-
-    /// Cast an [`f32`] to a [`real`][type@real] using `as`.
-    fn from_f32(f: f32) -> Self;
-
-    /// Cast an [`f64`] to a [`real`][type@real] using `as`.
-    fn from_f64(f: f64) -> Self;
-}
-
-#[cfg(not(feature = "double-precision"))]
-mod real_mod {
-    /// Floating point type used for many structs and functions in Godot.
-    ///
-    /// This type is `f32` by default, and `f64` when the Cargo feature `double-precision` is enabled.
-    ///
-    /// This is not the `float` type in GDScript; that type is always 64-bits. Rather, many structs in Godot may use
-    /// either 32-bit or 64-bit floats, for example [`Vector2`](super::Vector2). To convert between [`real`] and [`f32`] or
-    /// [`f64`], see [`RealConv`](super::RealConv).
-    ///
-    /// See also the [Godot docs on float](https://docs.godotengine.org/en/stable/classes/class_float.html).
-    // As this is a scalar value, we will use a non-standard type name.
-    #[allow(non_camel_case_types)]
-    pub type real = f32;
-
-    impl super::RealConv for real {
-        #[inline]
-        fn as_f32(self) -> f32 {
-            self
-        }
-
-        #[inline]
-        fn as_f64(self) -> f64 {
-            self as f64
-        }
-
-        #[inline]
-        fn from_f32(f: f32) -> Self {
-            f
-        }
-
-        #[inline]
-        fn from_f64(f: f64) -> Self {
-            f as f32
-        }
-    }
-
-    /// Re-export of [`std::f32::consts`] or [`std::f64::consts`], depending on precision config.
-    pub mod real_consts {
-        pub use std::f32::consts::*;
-    }
-
-    /// A 2-dimensional vector from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RVec2 = glam::Vec2;
-    /// A 3-dimensional vector from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RVec3 = glam::Vec3;
-    /// A 4-dimensional vector from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RVec4 = glam::Vec4;
-
-    /// A 2x2 column-major matrix from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RMat2 = glam::Mat2;
-    /// A 3x3 column-major matrix from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RMat3 = glam::Mat3;
-    /// A 4x4 column-major matrix from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RMat4 = glam::Mat4;
-
-    /// A matrix from [`glam`] quaternion representing an orientation. Using a floating-point format compatible
-    /// with [`real`].
-    pub type RQuat = glam::Quat;
-
-    /// A 2D affine transform from [`glam`], which can represent translation, rotation, scaling and
-    /// shear. Using a floating-point format compatible with [`real`].
-    pub type RAffine2 = glam::Affine2;
-    /// A 3D affine transform from [`glam`], which can represent translation, rotation, scaling and
-    /// shear. Using a floating-point format compatible with [`real`].
-    pub type RAffine3 = glam::Affine3A;
-}
-
-#[cfg(feature = "double-precision")]
-mod real_mod {
-    /// Floating point type used for many structs and functions in Godot.
-    ///
-    /// This type is `f32` by default, and `f64` when the Cargo feature `double-precision` is enabled.
-    ///
-    /// This is not the `float` type in GDScript; that type is always 64-bits. Rather, many structs in Godot may use
-    /// either 32-bit or 64-bit floats, for example [`Vector2`](super::Vector2). To convert between [`real`] and [`f32`] or
-    /// [`f64`], see [`RealConv`](super::RealConv).
-    ///
-    /// See also the [Godot docs on float](https://docs.godotengine.org/en/stable/classes/class_float.html).
-    ///
-    /// _Godot equivalent: `real_t`_
-    // As this is a scalar value, we will use a non-standard type name.
-    #[allow(non_camel_case_types)]
-    pub type real = f64;
-
-    impl super::RealConv for real {
-        #[inline]
-        fn as_f32(self) -> f32 {
-            self as f32
-        }
-
-        #[inline]
-        fn as_f64(self) -> f64 {
-            self
-        }
-
-        #[inline]
-        fn from_f32(f: f32) -> Self {
-            f as f64
-        }
-
-        #[inline]
-        fn from_f64(f: f64) -> Self {
-            f
-        }
-    }
-
-    /// Re-export of [`std::f32::consts`] or [`std::f64::consts`], depending on precision config.
-    pub mod real_consts {
-        pub use std::f64::consts::*;
-    }
-
-    /// A 2-dimensional vector from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RVec2 = glam::DVec2;
-    /// A 3-dimensional vector from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RVec3 = glam::DVec3;
-    /// A 4-dimensional vector from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RVec4 = glam::DVec4;
-
-    /// A 2x2 column-major matrix from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RMat2 = glam::DMat2;
-    /// A 3x3 column-major matrix from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RMat3 = glam::DMat3;
-    /// A 4x4 column-major matrix from [`glam`]. Using a floating-point format compatible with [`real`].
-    pub type RMat4 = glam::DMat4;
-
-    /// A matrix from [`glam`] quaternion representing an orientation. Using a floating-point format
-    /// compatible with [`real`].
-    pub type RQuat = glam::DQuat;
-
-    /// A 2D affine transform from [`glam`], which can represent translation, rotation, scaling and
-    /// shear. Using a floating-point format compatible with [`real`].
-    pub type RAffine2 = glam::DAffine2;
-    /// A 3D affine transform from [`glam`], which can represent translation, rotation, scaling and
-    /// shear. Using a floating-point format compatible with [`real`].
-    pub type RAffine3 = glam::DAffine3;
-}
-
-pub use crate::real;
-pub(crate) use real_mod::*;
-
-pub use real_mod::{real, real_consts};
-
-pub(crate) use glam::{IVec2, IVec3, IVec4};
-
-/// A macro to coerce float-literals into the [`real`] type.
-///
-/// Mainly used where you'd normally use a suffix to specify the type, such as `115.0f32`.
-///
-/// # Examples
-///
-/// Rust is not able to infer the `self` type of this call to `to_radians`:
-/// ```compile_fail
-/// use godot::builtin::real;
-///
-/// let radians: real = 115.0.to_radians();
-/// ```
-/// But we cannot add a suffix to the literal, since it may be either `f32` or
-/// `f64` depending on the context. So instead we use our macro:
-/// ```
-/// use godot::builtin::real;
-///
-/// let radians: real = real!(115.0).to_radians();
-/// ```
-#[macro_export]
-macro_rules! real {
-    ($f:literal) => {{
-        let f: $crate::builtin::real = $f;
-        f
-    }};
 }
 
 /// The side of a [`Rect2`] or [`Rect2i`].
